+++ library/Exceptions/NotBlankException.php
@@ -1,0 +1,34 @@
+<?php
+
+/*
+ * This file is part of Respect/Validation.
+ *
+ * (c) Alexandre Gomes Gaigalas <alexandre@gaigalas.net>
+ *
+ * For the full copyright and license information, please view the "LICENSE.md"
+ * file that was distributed with this source code.
+ */
+
+namespace Respect\Validation\Exceptions;
+
+class NotBlankException extends ValidationException
+{
+    const STANDARD = 0;
+    const NAMED = 1;
+
+    public static $defaultTemplates = array(
+        self::MODE_DEFAULT => array(
+            self::STANDARD => 'The value must not be blank',
+            self::NAMED => '{{name}} must not be blank',
+        ),
+        self::MODE_NEGATIVE => array(
+            self::STANDARD => 'The value must be blank',
+            self::NAMED => '{{name}} must be blank',
+        ),
+    );
+
+    public function chooseTemplate()
+    {
+        return $this->getName() == '' ? static::STANDARD : static::NAMED;
+    }
+}