+++ library/Validator.php
@@ -87,6 +87,7 @@ use Respect\Validation\Rules\Key;
  * @method static Validator no($useLocale = false)
  * @method static Validator noneOf()
  * @method static Validator not(Validatable $rule)
+ * @method static Validator notBlank()
  * @method static Validator notEmpty()
  * @method static Validator noWhitespace()
  * @method static Validator nullType()